requirements.txt
Flask==2.1.2
Flask-Limiter==2.8.0
passlib==1.7.4
pycryptodome==3.10.1
PyJWT==2.10.1
